move IsImageManifest to descriptor package
