+++ internal/descriptor/descriptor.go
@@ -1,0 +1,27 @@
+/*
+Copyright The ORAS Authors.
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+
+http://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.
+*/
+
+package descriptor
+
+import (
+	ocispec "github.com/opencontainers/image-spec/specs-go/v1"
+
+	"oras.land/oras/internal/docker"
+)
+
+// IsImageManifest checks whether a manifest is an image manifest.
+func IsImageManifest(desc ocispec.Descriptor) bool {
+	return desc.MediaType == docker.MediaTypeManifest || desc.MediaType == ocispec.MediaTypeImageManifest
+}
